requirements.txt
# Install necessary libraries
streamlit
streamlit-option-menu
langchain
langchain-groq
langchain-community
langchain_experimental
PyPDF2
faiss-cpu
beautifulsoup4
youtube-transcript-api
InstructorEmbedding
sentence-transformers
python-dotenv
tabulate
mysql-connector-python
google-generativeai
langchain-google-genai